Move Sidebar from App to Calendar and Map Page
