+++ utmap-client/src/components/pages/CalendarPage.js
@@ -5,14 +5,19 @@ import AddIcon from '@material-ui/icons/Add';
 import Dialog from '@material-ui/core/Dialog';
 import { ViewState } from '@devexpress/dx-react-scheduler';
 import {
-  Scheduler,
-  MonthView,
-  Toolbar,
-  DateNavigator,
-  Appointments,
-  TodayButton,
+	Scheduler,
+	MonthView,
+	Toolbar,
+	DateNavigator,
+	Appointments,
+	TodayButton,
 } from '@devexpress/dx-react-scheduler-material-ui';
 import CreateEventPage from './CreateEventPage';
+import { makeStyles } from '@material-ui/core/styles';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
+import SideBar from "./SideBar";
+import clsx from 'clsx';
 
 //example events REMOVE once the backend is connected
 const exampleEvents = [
@@ -21,7 +26,12 @@ const exampleEvents = [
     { startDate: '2021-03-27T12:00', endDate: '2021-03-30T13:30', title: 'Get jacked' }
 ]
 
-//Scheduler is the calendar, today, and taskbar components
+const useStyles = makeStyles(theme => ({
+    hide: {
+      display: 'none',
+    },
+  }));
+
 function CalendarPage() {
 	const [openEventForm, setOpenEventForm] = useState(false);
 	const [eventList, setEventList] = useState(exampleEvents);
@@ -36,18 +46,47 @@ function CalendarPage() {
 		setEventList([...eventList, eventData]);
 	}, [eventList]);
 
+	// Siderbar functions and variables
+	const classes = useStyles();
+    const [openDrawer, setOpenDrawer] = useState(false);
+
+    const handleDrawerOpen = () => {
+		setOpenDrawer(true);
+    };
+  
+    const handleDrawerClose = () => {
+		setOpenDrawer(false);
+    };
+
 	return (
 		<>
-		  <Paper>
-		    <Scheduler data={eventList}>
-		    <ViewState />
-		    <MonthView />
-		    <Toolbar />
-		    <DateNavigator />
-		    <TodayButton />
-		    <Appointments />
-		    </Scheduler>
-		  </Paper>
+			{/* Sidebar */}
+			<div align="right">
+				<IconButton
+					color="inherit"
+					aria-label="open drawer"
+					edge="end"
+					onClick={handleDrawerOpen}
+					className={clsx(openDrawer && classes.hide)}
+					alignment="right"
+					>
+					<MenuIcon />
+				</IconButton>
+				<SideBar open={openDrawer} onClose={handleDrawerClose}></SideBar>
+			</div>
+			{/* Calendar */}
+			<Paper>
+				<Scheduler data={eventList}>
+				<ViewState 
+					defaultCurrentViewName="Month"
+				/>
+				<MonthView />
+				<Toolbar />
+				<DateNavigator />
+				<TodayButton />
+				<Appointments />
+				</Scheduler>
+		  	</Paper>
 			{/* Add event button */}
 			<Button
 				variant='contained'
